Ptw/Add Form macro (#1)
Add a macro to generate a struct to represent a form. Handles creating a nested Form and Errors struct within a module, as well as creating functions for the field names. Updates all of the forms to use the associated field name functions

* get first form_struct working

* Handle trailing commas

* Add form and error objects

* Switch to using named errors

could later have a way to automatically generate conversions

* Move form_struct to its own file

And move specific interactions to html file for better locality.
Additionally, add macro_export instead of pub use.

* Replace form name's with field names from macro

Required adding delete contact list to macro,
which required adding support for field macros.
And reformat macro for easier reading.

* Update hx_triggers macro to be more readable too