Ptw/Embed contact attribute struct rather than duplicate (#2)
The duplication in our `model.rs` file for the different uses for contacts was fine, but wasn't the best to me. I wanted to see if another approach would work besides parameterizing `Contact` on whether it had an `ID` or not (I tried this before and couldn't get Diesel not to error when there wasn't an ID). The approach in this PR is how I would do it in golang (struct embedding), so this tries it out! It requires a specific, non-2.2.x version of Diesel to do this, but it seems more amenable to how other people would want to code since it avoids duplication.* get first form_struct working

* Try to embed shared data in a struct

but running into errors around:

```
error[E0412]: cannot find type `attributes` in module `crate::schema::contacts`
  --> src/model.rs:41:9
   |
41 |     pub attributes: ContactAttributes,
   |         ^^^^^^^^^^ not found in `crate::schema::contacts`
```

* Move from nightly to stable rust

* Move to git revision of Diesel

To include https://github.com/diesel-rs/diesel/pull/4209.
And implement deref for ease of use.